refactor: resolve code review P0/P1 issues — decouple settings循环依赖, split RequestContext, structured errors, frontend SWR cache
完成项:
- P0: settings_service 改用 broadcast channel 解耦 gateway_service 循环依赖
- P0: RequestContext 拆分为 RequestContextInner + Deref 代理，消除40行clone样板
- P1: ServiceError 改为 struct 变体，支持 code/detail 结构化错误
- P1: fetch_models 抽取为独立 model_fetcher 模块
- P1: 前端 providers/routingRules/groups 三个 Store 添加 30s SWR 缓存
- P1: DashboardView 添加跨 Store 联动自动刷新
- 其他: adapter_registry 延迟初始化, flush_batch 空值保护, 代码格式化

+++ src-tauri/src/persistence/gateway_key_repo.rs
@@ -18,7 +18,9 @@ impl GatewayKeyRepo {
         let enabled = if new.enabled.unwrap_or(true) { 1 } else { 0 };
         let key_hash = hash_api_key(&new.key_value);
         let key_prefix = format!("sk-gw-{}", &new.key_value[..8.min(new.key_value.len())]);
-        let max_concurrent = new.max_concurrent.unwrap_or(defaults::DEFAULT_KEY_MAX_CONCURRENT as u32);
+        let max_concurrent = new
+            .max_concurrent
+            .unwrap_or(defaults::DEFAULT_KEY_MAX_CONCURRENT);
 
         let row = sqlx::query(
             r#"
